n2t: has: Add symbols to the table correctly
Initializing the instruction number to -1 and then explicitly converting
it from s16 to u16 didn't seem very clean to me.

This bugged me so much that I fixed it to start the instruction number
from 0 and in the process discovered that there can be multiple labels
declared consecutively which would, then, all need to be translated to
the same instruction number.

This commit also passes all the available tests and `test.sh` passes
too.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>