n2t: vmt: Decouple push constant handling
Currently, `push constant` is handled in a unified manner, which makes
it non-ideal to implement push for other segments such as local, arg,
etc.

Separate out the code for generating `constant` segment from the `push`
code to allow for handling other segments as well. After all, push just
expects that the data that has to be pushed should be put in D register.

Also add phrase for the `local` segment.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>