n2t: vmt: Remove "FILE" from phrases except STATIC
The guidance in the slides is misleading - it asks to prefix function
and label names with filename whereas VM files have that prefixed
already. This leads to the special function "Sys.init" being rendered in
asm as "Sys.Sys.init" which leads to the emulator failing the test.

Remove "FILE" prefix from all phrases except STATIC and STATIC_ADDRESS.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>