n2t: vmt: Generalize intra-phrase labels
Add the "FILE" placeholder to the labels used within phrases so that
they may be safely used in multiple files that reside in a single
directory. Otherwise, the intra-phrase labels would not be unique,
leading to the assembler resolving the label in one file to its address
in another file when evaluating the result of the combined asm file from
multiple vm files.

Signed-off-by: Guru Das Srinagesh <gurooodas@gmail.com>